rename version file in ingest client
