+++ azure-kusto-ingest/azure/kusto/ingest/_version.py
